pyproject.toml
[tool.poetry]
name = "nonebot-plugin-tea-silencer"
version = "1.0.2"
description = "又一个消息审查插件，可针对不同程度嘴臭配置自动屏蔽、回应规则，支持指令拉黑群或用户。"
authors = ["悠岚岸"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_tea_silencer"}]
homepage = "https://github.com/youlanan/nonebot-plugin-tea-silencer"
repository = "https://github.com/youlanan/nonebot-plugin-tea-silencer"
keywords = ["NoneBot", "silencer"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.7.1"
nonebot2 = "^2.2.1"
nonebot-adapter-onebot = "^2.2.0"
nonebot_plugin_localstore = ">=0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"